product_info/requirement.txt
mysql-connector-python
selenium
